refactor(assets): consolidate tailwind/ and components/ into one flat CSS tree
Merges the two CSS trees that grew out of Phase 2's migration history
(tailwind/*.css, the Bootstrap/Limitless port; components/*.css, this
app's own hand-written CSS) into one feature-grouped set under
assets/styles/, and a single global entry (app.css replaces app.css +
tailwind.css). That split was an accident of migration history, not a
real architecture boundary, and it hid real duplication: colors and
breakpoints defined once per side instead of once, and .list-feed
outright diverging between the two.

First step of captain-coaster/captain-coaster#385's now-unified scope
(consolidate -> tokens/@layer -> font-size -> grid -> Preflight ->
robustness pass), landing as one PR with ordered commits rather than
separate PRs.

Behavior-preserving except three verified, called-out fixes:
- .list-feed: panels-typography.css and timeline.css each defined the
  container padding differently (7px vs 10px); consolidated on 7px,
  the value actually winning in the current build.
- forms.css's own font-size breakpoint drifted to 768px while every
  other breakpoint in the app is 769px; aligned to 769px.
- icons.css's .w-4../.h-4.. rules were byte-identical duplicates of
  Tailwind's own default utilities, unlayered so they always won for
  no actual value; removed, real SVG-spacing rules kept.

Also drops two confirmed-redundant CSS imports (verified against real
template/controller usage, not assumed): rating-actions.css (now part
of the global rating.css) was also imported per-entry from coaster.js;
top-list.css was imported both globally and per-entry even though its
classes only ever render on Top/edit.html.twig.

Every merged file's declarations were diffed against its source
files' to confirm zero unintended content loss. Verified live via
Playwright (desktop + 390x844 mobile): homepage, ranking pagination,
coaster show page, account-menu dropdown, mobile nav -- all pixel-
identical to pre-change screenshots.
