+++ assets/styles/alerts.css
@@ -1,0 +1,140 @@
+/*
+ * .alert — migrated from theme/bootstrap-limitless/alerts.less +
+ * bootstrap/less/alerts.less. Values below are copied from the compiled
+ * output of those two files (cascade already resolved), not re-derived
+ * from LESS variables, to guarantee an iso-visual result. Written as plain
+ * CSS rather than @apply: several rules here are pseudo-element icons
+ * (base64 SVG backgrounds) and precise absolute positioning that gain
+ * nothing from utility composition and are safer to keep as a verbatim
+ * copy of known-correct values.
+ *
+ * Only variants actually reachable in this app are included: alert-success/
+ * -danger/-warning/-info (all reachable via addFlash()), plus
+ * alert-styled-left/alert-arrow-left/alert-dismissible/alert-component/
+ * alert-heading (the only *styled* modifiers referenced in templates).
+ * "alert-bordered" is also referenced in templates (flashes.html.twig,
+ * connect/login.html.twig, registration/register.html.twig) but was never
+ * defined in the original LESS either -- pre-existing dead weight, not
+ * something this migration should invent styling for.
+ * alert-primary and the .ui-pnotify integration were never used and are
+ * dropped. "alert-error" (also a real addFlash() value) intentionally gets
+ * no color rule below, matching the current (accidental) unstyled look —
+ * "alert-error" was never a real Bootstrap/Limitless variant.
+ */
+.alert {
+    position: relative;
+    margin-bottom: 20px;
+    padding: 15px 20px;
+    border: 1px solid transparent;
+    border-radius: 3px;
+}
+
+.alert > p,
+.alert > ul {
+    margin-bottom: 0;
+}
+
+.alert > p + p {
+    margin-top: 5px;
+}
+
+.alert-heading {
+    margin-top: 0;
+    margin-bottom: 5px;
+}
+
+.alert .close {
+    color: inherit;
+}
+
+.alert-dismissible,
+.alert-dismissable {
+    padding-right: 35px;
+}
+
+.alert-dismissible .close,
+.alert-dismissable .close {
+    position: relative;
+    top: -2px;
+    right: -21px;
+}
+
+.alert-success {
+    color: #205823;
+    background-color: #e8f5e9;
+    border-color: #4caf50;
+}
+
+.alert-info {
+    color: #00545c;
+    background-color: #e0f7fa;
+    border-color: #00bcd4;
+}
+
+.alert-warning {
+    color: #aa3510;
+    background-color: #fff3e0;
+    border-color: #ff9800;
+}
+
+.alert-danger {
+    color: #9c1f1f;
+    background-color: #fbe9e7;
+    border-color: #ff5722;
+}
+
+/* Only relevant paired with alert-styled-left, per the templates that use it. */
+.alert-component[class*="alert-styled-"] {
+    background-color: #fff;
+}
+
+.alert-styled-left {
+    border-left-width: 44px;
+}
+
+/* Decorative icon in the colored left border of a styled alert. */
+.alert[class*="alert-styled-"]:after {
+    content: "";
+    display: inline-block;
+    position: absolute;
+    top: 50%;
+    left: -44px;
+    width: 44px;
+    height: 16px;
+    margin-top: -8px;
+    line-height: 1;
+    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGZpbGw9Im5vbmUiIHZpZXdCb3g9IjAgMCAyNCAyNCIgc3Ryb2tlLXdpZHRoPSIxLjUiIHN0cm9rZT0iY3VycmVudENvbG9yIiBhcmlhLWhpZGRlbj0idHJ1ZSIgZGF0YS1zbG90PSJpY29uIj48cGF0aCBzdHJva2UtbGluZWNhcD0icm91bmQiIHN0cm9rZS1saW5lam9pbj0icm91bmQiIGQ9Im0xMS4yNSAxMS4yNS4wNDEtLjAyYS43NS43NSAwIDAgMSAxLjA2My44NTJsLS43MDggMi44MzZhLjc1Ljc1IDAgMCAwIDEuMDYzLjg1M2wuMDQxLS4wMjFNMjEgMTJhOSA5IDAgMSAxLTE4IDAgOSA5IDAgMCAxIDE4IDBabS05LTMuNzVoLjAwOHYuMDA4SDEyVjguMjVaIi8+PC9zdmc+);
+    background-position: 50%;
+    background-repeat: no-repeat;
+    background-size: contain;
+}
+
+.alert[class*="alert-styled-"].alert-danger:after {
+    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCAyNCAyNCIgZmlsbD0iY3VycmVudENvbG9yIiBhcmlhLWhpZGRlbj0idHJ1ZSIgZGF0YS1zbG90PSJpY29uIj48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGQ9Ik0xMiAyLjI1Yy01LjM4NSAwLTkuNzUgNC4zNjUtOS43NSA5Ljc1czQuMzY1IDkuNzUgOS43NSA5Ljc1IDkuNzUtNC4zNjUgOS43NS05Ljc1UzE3LjM4NSAyLjI1IDEyIDIuMjVabS0xLjcyIDYuOTdhLjc1Ljc1IDAgMSAwLTEuMDYgMS4wNkwxMC45NCAxMmwtMS43MiAxLjcyYS43NS43NSAwIDEgMCAxLjA2IDEuMDZMMTIgMTMuMDZsMS43MiAxLjcyYS43NS43NSAwIDEgMCAxLjA2LTEuMDZMMTMuMDYgMTJsMS43Mi0xLjcyYS43NS43NSAwIDEgMC0xLjA2LTEuMDZMMTIgMTAuOTRsLTEuNzItMS43MloiIGNsaXAtcnVsZT0iZXZlbm9kZCIvPjwvc3ZnPg==);
+}
+
+.alert[class*="alert-styled-"].alert-success:after {
+    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCAyNCAyNCIgZmlsbD0iY3VycmVudENvbG9yIiBhcmlhLWhpZGRlbj0idHJ1ZSIgZGF0YS1zbG90PSJpY29uIj48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGQ9Ik0yLjI1IDEyYzAtNS4zODUgNC4zNjUtOS43NSA5Ljc1LTkuNzVzOS43NSA0LjM2NSA5Ljc1IDkuNzUtNC4zNjUgOS43NS05Ljc1IDkuNzVTMi4yNSAxNy4zODUgMi4yNSAxMlptMTMuMzYtMS44MTRhLjc1Ljc1IDAgMSAwLTEuMjItLjg3MmwtMy4yMzYgNC41M0w5LjUzIDEyLjIyYS43NS43NSAwIDAgMC0xLjA2IDEuMDZsMi4yNSAyLjI1YS43NS43NSAwIDAgMCAxLjE0LS4wOTRsMy43NS01LjI1WiIgY2xpcC1ydWxlPSJldmVub2RkIi8+PC9zdmc+);
+}
+
+.alert[class*="alert-styled-"].alert-warning:after {
+    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGZpbGw9Im5vbmUiIHZpZXdCb3g9IjAgMCAyNCAyNCIgc3Ryb2tlLXdpZHRoPSIxLjUiIHN0cm9rZT0iY3VycmVudENvbG9yIiBhcmlhLWhpZGRlbj0idHJ1ZSIgZGF0YS1zbG90PSJpY29uIj48cGF0aCBzdHJva2UtbGluZWNhcD0icm91bmQiIHN0cm9rZS1saW5lam9pbj0icm91bmQiIGQ9Ik0xMiA5djMuNzVtLTkuMzAzIDMuMzc2Yy0uODY2IDEuNS4yMTcgMy4zNzQgMS45NDggMy4zNzRoMTQuNzFjMS43MyAwIDIuODEzLTEuODc0IDEuOTQ4LTMuMzc0TDEzLjk0OSAzLjM3OGMtLjg2Ni0xLjUtMy4wMzItMS41LTMuODk4IDBMMi42OTcgMTYuMTI2Wk0xMiAxNS43NWguMDA3di4wMDhIMTJ2LS4wMDhaIi8+PC9zdmc+);
+}
+
+.alert[class*="alert-styled-"].alert-info:after {
+    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGZpbGw9Im5vbmUiIHZpZXdCb3g9IjAgMCAyNCAyNCIgc3Ryb2tlLXdpZHRoPSIxLjUiIHN0cm9rZT0iY3VycmVudENvbG9yIiBhcmlhLWhpZGRlbj0idHJ1ZSIgZGF0YS1zbG90PSJpY29uIj48cGF0aCBzdHJva2UtbGluZWNhcD0icm91bmQiIHN0cm9rZS1saW5lam9pbj0icm91bmQiIGQ9Im0xMS4yNSAxMS4yNS4wNDEtLjAyYS43NS43NSAwIDAgMSAxLjA2My44NTJsLS43MDggMi44MzZhLjc1Ljc1IDAgMCAwIDEuMDYzLjg1M2wuMDQxLS4wMjFNMjEgMTJhOSA5IDAgMSAxLTE4IDAgOSA5IDAgMCAxIDE4IDBabS05LTMuNzVoLjAwOHYuMDA4SDEyVjguMjVaIi8+PC9zdmc+);
+}
+
+/* Small triangle notch at the edge of the colored left border. */
+.alert[class*="alert-arrow-"]:before {
+    content: "";
+    display: inline-block;
+    position: absolute;
+    top: 50%;
+    left: 0;
+    margin-top: -5px;
+    border-top: 5px solid transparent;
+    border-bottom: 5px solid transparent;
+    border-left: 5px solid;
+    border-left-color: inherit;
+}
