--- assets/styles/components/enhanced-select.css
@@ -1,69 +1,0 @@
-/*
- * Tom Select overrides -- restyles its own `.ts-*` classes (library CSS
- * imported by enhanced_select_controller.js) to match this app's form
- * controls instead of Tom Select's default look.
- */
-/*
- * Tom Select copies the original <select>'s classes (including
- * `form-control`) onto `.ts-wrapper`. `.form-control`'s fixed `height: 36px`
- * (forms-buttons.css) then clips the wrapper when selected tags wrap to a
- * second line -- the overflow stays visible and bleeds into the next form
- * field instead of growing the box. Override it back to auto-height here.
- */
-/*
- * `.form-control` also draws its own border/padding/background -- since
- * `.ts-wrapper` carries that class too, without this it nests a second box
- * around `.ts-control` below (visible as a double border). Neutralize the
- * outer box entirely; `.ts-control` is the only one meant to be visible.
- */
-.ts-wrapper.form-control {
-    height: auto;
-    min-height: 44px;
-    padding: 0;
-    border: none;
-    background: none;
-}
-
-.ts-wrapper .ts-control {
-    min-height: 44px;
-    padding: 6px 12px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    background-color: #fff;
-}
-
-.ts-wrapper.focus .ts-control {
-    border-color: #2196f3;
-    box-shadow: 0 0 0 2px rgba(33, 150, 243, 0.2);
-}
-
-/*
- * Tom Select's own equivalent rules (`.ts-wrapper.multi .ts-control > div`,
- * `.ts-dropdown .active`) tie or beat these on specificity and load after
- * this file (their CSS is bundled with the lazy-loaded controller chunk,
- * injected later at runtime) -- the extra `.ts-control`/`.ts-dropdown-
- * content` qualifiers below aren't decorative, they're what makes these
- * actually win instead of losing the cascade to load order.
- */
-.ts-wrapper.plugin-remove_button .ts-control .item {
-    background: #2196f3;
-    color: #fff;
-    border: none;
-    border-radius: 3px;
-    font-size: 12px;
-}
-
-.ts-wrapper.plugin-remove_button .ts-control .item .remove {
-    border-left-color: rgba(255, 255, 255, 0.3);
-}
-
-.ts-dropdown {
-    border: 1px solid #ccc;
-    border-radius: 0 0 4px 4px;
-    font-size: 13px;
-}
-
-.ts-dropdown .ts-dropdown-content .active {
-    background-color: #2196f3;
-    color: #fff;
-}
