--- assets/styles/components/forms.css
@@ -1,91 +1,0 @@
-/* ------------------------------------------------------------------------------
- *
- *  # Global Form Styles
- *
- *  Global form input styles to prevent mobile auto-zoom and ensure consistent sizing
- *  Mobile devices zoom when input font-size is < 16px, so we set 16px as minimum
- *  Desktop uses media queries to reduce to more appropriate sizes
- *
- * ---------------------------------------------------------------------------- */
-
-/* Base styles for all form inputs - prevents mobile zoom */
-input[type="text"],
-input[type="email"],
-input[type="password"],
-input[type="search"],
-input[type="tel"],
-input[type="url"],
-input[type="number"],
-input[type="date"],
-input[type="datetime-local"],
-input[type="month"],
-input[type="week"],
-input[type="time"],
-select,
-textarea {
-    font-size: 16px !important; /* Minimum 16px prevents iOS zoom */
-    min-height: 44px; /* Better touch targets for mobile users */
-}
-
-/* Date input specific fixes */
-input[type="date"] {
-    /* Fix vertical alignment on mobile - center the text */
-    line-height: normal;
-    cursor: pointer;
-}
-
-/* Fix for webkit browsers (Safari, Chrome) date input alignment */
-input[type="date"]::-webkit-date-and-time-value {
-    text-align: left;
-}
-
-/*
- * Make calendar icon more visible. Sized explicitly rather than left to
- * scale with font-size: the 16px font forced above (anti-zoom) would
- * otherwise inflate this icon noticeably larger than the rest of the
- * field's chrome.
- */
-input[type="date"]::-webkit-calendar-picker-indicator {
-    cursor: pointer;
-    opacity: 0.6;
-    width: 16px;
-    height: 16px;
-}
-
-input[type="date"]::-webkit-calendar-picker-indicator:hover {
-    opacity: 1;
-}
-
-/*
- * `.datepicker-input` (added by datepicker_controller.js to the review
- * form's ridden-date field) still gets font-size:16px!important from the
- * base rule above -- can't go lower, iOS zooms the page on focus below
- * that. Trims the surrounding box instead so it doesn't look oversized
- * next to this page's other .form-control fields, mirroring the same
- * "shrink everything but the font" fix applied to .small-date-input
- * (assets/styles/components/rating-actions.css).
- */
-.datepicker-input {
-    min-height: 40px;
-    padding: 6px 10px;
-}
-
-/* Desktop: reduce font size for better proportions (768px and up) */
-@media (min-width: 768px) {
-    input[type="text"],
-    input[type="email"],
-    input[type="password"],
-    input[type="search"],
-    input[type="tel"],
-    input[type="url"],
-    input[type="number"],
-    input[type="date"],
-    input[type="datetime-local"],
-    input[type="month"],
-    input[type="week"],
-    input[type="time"],
-    select,
-    textarea {
-        font-size: 13px !important;
-    }
-}
