--- assets/styles/components/notification-center.css
@@ -1,52 +1,0 @@
-/* Converted from LESS to plain CSS: @text-color -> #333, @color-primary-600
- * -> #1e88e5, @color-primary-500 -> #2196f3, fade(@text-color,45%) ->
- * rgba(51,51,51,.45).
- *
- * Read/unread needs to register at a glance across a long list without
- * shouting on every row. A border-left accent + background tint (the
- * default reach for "unread row") reads as a solid block of color per
- * item and gets loud fast in a list of 15-20 - the treatment below instead
- * de-emphasizes read rows (muted text, faded icon) and marks unread with
- * weight plus a small dot, the same shape Gmail/GitHub/Linear converge on.
- */
-.notif-item .media-body a {
-    /* Overrides the theme's default link-blue: the message is content,
-     * not a "this is a link" affordance - only the hover/focus state
-     * below needs to signal that.
-     */
-    color: #333;
-}
-
-.notif-item .media-body a:hover,
-.notif-item .media-body a:focus {
-    color: #1e88e5;
-}
-
-.notif-item.unread .media-body a {
-    font-weight: 600;
-}
-
-.notif-item.unread .media-annotation {
-    position: relative;
-    padding-left: 11px;
-}
-
-.notif-item.unread .media-annotation::before {
-    content: "";
-    position: absolute;
-    left: 0;
-    top: 50%;
-    width: 6px;
-    height: 6px;
-    margin-top: -3px;
-    border-radius: 50%;
-    background-color: #2196f3;
-}
-
-.notif-item:not(.unread) .media-left .btn {
-    opacity: 0.5;
-}
-
-.notif-item:not(.unread) .media-annotation {
-    color: rgba(51, 51, 51, 0.45);
-}
