--- assets/styles/icons/icons.css
@@ -1,94 +1,0 @@
-/* Icon size utility classes for Heroicons */
-
-/* Width utilities */
-.w-4 {
-    width: 1rem; /* 16px */
-}
-
-.w-5 {
-    width: 1.25rem; /* 20px */
-}
-
-.w-6 {
-    width: 1.5rem; /* 24px */
-}
-
-.w-8 {
-    width: 2rem; /* 32px */
-}
-
-.w-10 {
-    width: 2.5rem; /* 40px */
-}
-
-.w-12 {
-    width: 3rem; /* 48px */
-}
-
-.w-16 {
-    width: 4rem; /* 64px */
-}
-
-/* Height utilities */
-.h-4 {
-    height: 1rem; /* 16px */
-}
-
-.h-5 {
-    height: 1.25rem; /* 20px */
-}
-
-.h-6 {
-    height: 1.5rem; /* 24px */
-}
-
-.h-8 {
-    height: 2rem; /* 32px */
-}
-
-.h-10 {
-    height: 2.5rem; /* 40px */
-}
-
-.h-12 {
-    height: 3rem; /* 48px */
-}
-
-.h-16 {
-    height: 4rem; /* 64px */
-}
-
-/* Ensure SVG icons display inline properly */
-svg.w-4,
-svg.w-5,
-svg.w-6,
-svg.w-8,
-svg.w-10,
-svg.w-12,
-svg.w-16 {
-    display: inline-block;
-    vertical-align: middle;
-}
-
-/* Icon spacing for navbar and sidebar */
-.navbar a svg {
-    margin-right: 0;
-    margin-left: 0;
-}
-
-/* Add spacing only for dropdown menu items with text */
-.navbar .dropdown-menu a svg {
-    margin-right: 8px;
-}
-
-.sidebar .navigation a svg {
-    margin-right: 8px;
-}
-
-.sidebar-xs .sidebar .navigation a svg {
-    margin-right: 0;
-}
-
-.list-group-item svg {
-    margin-right: 8px;
-}
