+++ assets/styles/layout.css
@@ -1,16 +1,9 @@
 /*
- * Layout -- migrated together: theme/core/layout/{content,footer,
+ * Page shell -- migrated together: theme/core/layout/{content,footer,
  * component-animation}.less. `sidebar.less` (also theme/core/layout) is
- * deliberately NOT part of this cluster despite living in the same
- * folder -- it drives real jQuery toggle behavior (`layout_fixed_custom.js`,
- * `theme.js`, the navbar's `sidebar-*-toggle` buttons) needing the same
- * kind of Stimulus rewrite the Navigation cluster did, so it's its own,
- * separately-scoped piece of this phase. Checked for selector coupling
- * with `.content-wrapper` first: sidebar.less's only reference is
- * `.sidebar-fixed-expanded .content-wrapper { padding-left: ... }`, a
- * compound selector that keeps winning over the bare `.content-wrapper`
- * base rule ported here regardless of cascade position (2 classes beats
- * 1), so migrating content.less first doesn't risk that override.
+ * deliberately its own file (sidebar.css) -- it drives real JS toggle
+ * behavior (sidebar_controller.js) needing its own scope, unrelated to
+ * this shell.
  *
  * `.page-container`/`.page-content`/`.content-wrapper`/`.content` are
  * base.html.twig's page shell -- rendered on every page, unconditional.
@@ -32,22 +25,13 @@
  *
  * `.footer`'s `&.navbar`/`-fixed-bottom` branch and `.footer-boxed` are
  * dropped: `footer.html.twig` never combines `.footer` with `.navbar`,
- * and `.footer-boxed`/`navbar-fixed-bottom` have zero real uses (the one
- * `$('.navbar-fixed-bottom')` query in theme.js always matches nothing).
+ * and `.footer-boxed`/`navbar-fixed-bottom` have zero real uses.
  *
  * component-animation.less scope trimmed to selectors real markup
  * actually has: `.btn`, `.navbar-nav > li > a`, `.dropdown-menu > li >
  * a`, `.list-group-item`, `.pagination > li > a/span`, `.close`, `.table
  * tr/th/td`, and `.no-transitions *` (toggled on `<body>` by theme.js
- * during resize). Everything else in that file's giant shared-declaration
- * selector list targets jQuery-plugin widgets this app never uses
- * (tokenfield, select2, dataTables, pikaday, spectrum, noUiSlider,
- * fullcalendar, plupload, jQuery UI, AnyTime) or already-dead classes
- * from earlier clusters (`.breadcrumb`, `.pager`, `.nav-tabs`/`.nav-pills`,
- * `a.label`/`a.badge` -- neither is ever an anchor per the Labels &
- * Badges audit, `.media-link`, `.navbar-link`, `.tag [data-role=remove]`,
- * `.dropdown-menu > li > label` -- no template puts a bare `<label>`
- * inside a `.dropdown-menu` item).
+ * during resize).
  */
 .page-container {
     position: relative;
@@ -120,4 +104,49 @@
 .table th,
 .table td {
     transition: background-color ease-in-out 0.15s;
+}
+
+/*
+ * Login page layout -- overrides the table-based shell above with
+ * flexbox for the one page (connect/login.html.twig) that centers a
+ * single form instead of running the sidebar/content-wrapper layout.
+ */
+body.login-container {
+    margin: 0;
+    padding: 0;
+    min-height: 100vh;
+    display: flex;
+}
+
+body.login-container .page-container {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    min-height: 100vh;
+}
+
+body.login-container .page-content {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+}
+
+body.login-container .content-wrapper {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+}
+
+body.login-container .content {
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 20px;
+}
+
+.login-form-container {
+    width: 100%;
+    max-width: 500px;
 }
