+++ assets/styles/navbar.css
@@ -16,8 +16,8 @@
  * apply cleanly here since navbar/dropdowns/navs/modals are all genuinely
  * used -- just far narrower than their raw line counts suggest.
  *
- * account-menu.less (assets/styles/components/, NOT part of this migration)
- * already owns all of the navbar account-menu's actual visual styling and
+ * account-menu.css (assets/styles/, not part of this migration) already
+ * owns all of the navbar account-menu's actual visual styling and
  * its own mobile dropdown override (bypasses .open entirely, renders the
  * account panel inline once #navbar-mobile is expanded) -- confirmed via its
  * own comments that it depends on @grid-float-breakpoint and
@@ -827,24 +827,19 @@
 }
 
 /*
- * account-menu.less (assets/styles/components/, not part of this migration)
- * relies on out-ranking two of the base rules above -- `.account-menu li > a
- * { display: flex }` (for the Language/Units rows' label-left, value-right
- * layout) and `.dropdown-submenu > a.account-menu-row:after { display: none }`
- * (to hide the submenu chevron this file draws via ::after, replaced by its
- * own chevron-right icon inside .account-menu-row-value). Both are tied in
- * specificity with the base rules here and used to win by simply coming
- * later in app.less's cascade. Now that this CSS lives in tailwind.css
- * (loaded after the *entire* app.less bundle), the tie flips the other way --
- * `.dropdown-menu > li > a { display: block }` and the chevron ::after above
- * started winning instead, breaking the account menu's row layout and
- * showing a stray black chevron on Language/Units. Re-asserted here so
- * order is deterministic regardless of account-menu.less's own position.
+ * account-menu.css relies on out-ranking two of the base rules above --
+ * `.account-menu li > a { display: flex }` (for the Language/Units rows'
+ * label-left, value-right layout) and `.dropdown-submenu > a.account-
+ * menu-row:after { display: none }` (to hide the submenu chevron this
+ * file draws via ::after, replaced by its own chevron-right icon inside
+ * .account-menu-row-value). Both used to be re-asserted here too, as a
+ * duplicate, because a Phase 2-era cascade-order flip (app.less's LESS
+ * bundle vs. tailwind.css loading after it) made this file's version
+ * win when the two disagreed. Since account-menu.css and this file now
+ * load from the same consolidated app.css, in the same relative order as
+ * before (account-menu.css first), and both copies agreed on the value
+ * regardless of which won, the duplicate was dead weight -- removed
+ * (captain-coaster/captain-coaster#385 consolidation pass). If this file
+ * ever needs its own `.account-menu`-context override again, put it in
+ * account-menu.css, not here.
  */
-.account-menu li > a {
-    display: flex;
-}
-
-.dropdown-submenu > a.account-menu-row:after {
-    display: none;
-}
