+++ assets/styles/responsive-visibility.css
@@ -1,0 +1,74 @@
+/*
+ * Responsive utilities -- migrated: bootstrap/less/responsive-utilities.less.
+ * Only the `xs` breakpoint (max-width:768px, matching every other
+ * breakpoint used throughout this migration) has real uses: `.hidden-xs`
+ * (rating_table.html.twig, ranking/info.html.twig, page_header.html.twig,
+ * navbar.html.twig), `.visible-xs` and `.visible-xs-block` (ranking/info,
+ * navbar). `.{visible,hidden}-{sm,md,lg}`, every `-inline`/`-inline-block`
+ * variant, and the print-only `.visible-print*`/`.hidden-print` are all
+ * confirmed dead (0 uses). `.responsive-visibility()`'s `table&`/`tr&`/
+ * `th&`/`td&` sub-selectors are also dropped -- every real use here is a
+ * plain `<span>`/`<ul>`, never a table element.
+ */
+@media (max-width: 768px) {
+    .hidden-xs {
+        display: none !important;
+    }
+}
+
+.visible-xs {
+    display: none !important;
+}
+
+.visible-xs-block {
+    display: none !important;
+}
+
+@media (max-width: 768px) {
+    .visible-xs {
+        display: block !important;
+    }
+
+    .visible-xs-block {
+        display: block !important;
+    }
+}
+
+/*
+ * Component animations -- migrated: bootstrap/less/component-animations.less.
+ * `.collapse`/`.collapse.in` below is exactly what the `.collapse{
+ * visibility:visible}` fix at the top of the entry stylesheet counters
+ * (Tailwind's own `.collapse` utility's `visibility:collapse`), while
+ * `display:none`/`.in{display:block}` (a different property, no
+ * conflict) always needed to come from *somewhere* unlayered; this file
+ * is that source.
+ *
+ * `.fade`/`.fade.in` is real and needed by the review-report modal
+ * (`class="modal fade"` in _review_item.html.twig, toggled by
+ * modal_controller.js) -- its own backdrop is now the browser's native
+ * `<dialog>` `::backdrop`, styled directly in navbar.css, no longer a
+ * `.fade`-toggled element of its own.
+ *
+ * `.collapse`/`.in` is real too (`#navbar-mobile`'s `class="navbar-collapse
+ * collapse"`, toggled via `.in` by navbar_collapse_controller.js).
+ * `.collapsing` and the `tr&.in`/`tbody&.in` table variants are NOT
+ * ported: no JS anywhere ever adds a `.collapsing` class
+ * (`navbar_collapse_controller.js` toggles `.in` directly), and no
+ * template collapses a `<tr>`/`<tbody>`.
+ */
+.fade {
+    opacity: 0;
+    transition: opacity 0.15s linear;
+}
+
+.fade.in {
+    opacity: 1;
+}
+
+.collapse {
+    display: none;
+}
+
+.collapse.in {
+    display: block;
+}
