--- assets/styles/tailwind/component-animations.css
@@ -1,41 +1,0 @@
-/*
- * Component animations -- migrated: bootstrap/less/component-animations.less.
- * `.collapse`/`.collapse.in` below is exactly the "Bootstrap's unrelated
- * `.collapse` JS-toggle class (component-animations.less)" the top-of-file
- * `.collapse{visibility:visible}` fix (PR #384) already refers to -- that
- * fix countered Tailwind's own `.collapse` utility's `visibility:collapse`,
- * while `display:none`/`.in{display:block}` (a different property, no
- * conflict) always needed to come from *somewhere* unlayered; this file is
- * now that source instead of the LESS bundle.
- *
- * `.fade`/`.fade.in` is real and needed by the review-report modal
- * (`class="modal fade"` in _review_item.html.twig, toggled by
- * modal_controller.js) -- its own backdrop is now the browser's native
- * `<dialog>` `::backdrop`, styled directly in navigation.css, no longer
- * a `.fade`-toggled element of its own.
- *
- * `.collapse`/`.in` is real too (`#navbar-mobile`'s `class="navbar-collapse
- * collapse"`, toggled via `.in` by navbar_collapse_controller.js).
- * `.collapsing` and the `tr&.in`/`tbody&.in` table variants are NOT
- * ported: no JS anywhere ever adds a `.collapsing` class (`navbar_collapse_
- * controller.js` toggles `.in` directly, matching Limitless's own now-dead
- * `.navbar-collapse.collapsing{transition-duration:.00000001s}` override,
- * which existed only to defeat an animation nothing ever triggers), and
- * no template collapses a `<tr>`/`<tbody>`.
- */
-.fade {
-    opacity: 0;
-    transition: opacity 0.15s linear;
-}
-
-.fade.in {
-    opacity: 1;
-}
-
-.collapse {
-    display: none;
-}
-
-.collapse.in {
-    display: block;
-}
