+++ assets/styles/youtube-thumbnail.css
